# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/manage/customers/{id}/profile": {
      "get": {
        "operationId": "ManageController_customerProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "Customer 360 — balance, tier, transactions, badges, referrals.",
        "tags": [
          "brand-admin"
        ]
      },
      "patch": {
        "operationId": "ManageController_updateCustomerProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "Update a customer profile (name, gender, birthdate).",
        "tags": [
          "brand-admin"
        ]
      }
    },
...
    "/customer/wallet/cards": {
      "get": {
        "operationId": "CustomerWalletController_cards",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every card in the wallet, with balance and tier progress.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/vouchers": {
      "get": {
        "operationId": "CustomerWalletController_vouchers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rewards held across all brands.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/activity": {
      "get": {
        "operationId": "CustomerWalletController_activity",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One timeline across every brand: points and reward events.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/profile": {
      "get": {
        "operationId": "CustomerWalletController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The signed-in person’s own details.",
        "tags": [
          "customer"
        ]
      },
      "patch": {
        "operationId": "CustomerWalletController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWalletProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit name, gender or birthdate.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/brands": {
      "get": {
        "operationId": "CustomerWalletController_brands",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Brands to discover and join.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/brands/{brandId}/join": {
      "post": {
        "operationId": "CustomerWalletController_join",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a brand from the app and get a card.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/cards/{brandId}/scan-code": {
      "get": {
        "operationId": "CustomerWalletController_scanCode",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The value to show as a QR at this brand’s till.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/cards/{brandId}/token": {
      "post": {
        "operationId": "CustomerWalletController_brandToken",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mint a brand-scoped customer token for one card.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/email": {
      "patch": {
        "operationId": "CustomerWalletController_setEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set or clear the signed-in person’s email address.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/push-token": {
      "post": {
        "operationId": "CustomerWalletController_registerPushToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register this device for notifications.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/offers": {
      "get": {
        "operationId": "CustomerWalletController_offers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Live campaigns at the brands this customer holds.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/branch-visits": {
      "get": {
        "operationId": "CustomerWalletController_branchVisits",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Branches this person visits, busiest first — the home-area suggestion.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/home-branch": {
      "patch": {
        "operationId": "CustomerWalletController_setHomeBranch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHomeBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set or clear the home area. Null clears it.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/sessions": {
      "get": {
        "operationId": "CustomerWalletController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Devices currently signed in to this wallet.",
        "tags": [
          "customer"
        ]
      }
    },
    "/customer/wallet/sessions/{id}": {
      "delete": {
        "operationId": "CustomerWalletController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign one other device out of this wallet.",
        "tags": [
          "customer"
        ]
      }
    },
...
  },
  "info": {
    "title": "RFM Loyalty Engine API",
    "description": "Multi-tenant, closed-loop, B2B2C loyalty engine. Four surfaces: Superadmin (/v1/admin), Brand Admin (/v1/manage), Customer (/v1/customer), Terminal (/v1/terminal).",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "customer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "terminal-hmac": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "UpdateCustomerProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "description": "e.g. male / female / other / undisclosed"
          },
          "birthdate": {
            "type": "object",
            "description": "ISO date (YYYY-MM-DD); empty string clears it"
          }
        }
      },
...
      "UpdateWalletProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Zain Ahmed"
          },
          "gender": {
            "type": "string",
            "example": "female"
          },
          "birthdate": {
            "type": "object",
            "example": "1994-03-17",
            "description": "ISO date; null clears it."
          },
          "nationality": {
            "type": "object",
            "example": "AE",
            "description": "ISO 3166-1 alpha-2 country code; null clears it."
          },
          "txnAlertsOptOut": {
            "type": "boolean",
            "description": "Stop the WhatsApp message sent after each transaction. Sign-in codes are unaffected."
          }
        }
      },
      "SetEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "object",
            "nullable": true,
            "description": "Email address, or null to clear."
          }
        }
      },
      "RegisterPushTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Expo push token for this install."
          },
          "platform": {
            "type": "string",
            "description": "ios | android"
          }
        }
      },
      "SetHomeBranchDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "object",
            "nullable": true,
            "description": "Branch id, or null to clear."
          }
        }
      }
    }
  }
}